% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fql-methods.R
\name{coef.fql}
\alias{coef.fql}
\title{Extract Coefficients from an FQL Object}
\usage{
\method{coef}{fql}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fql"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Named vector of estimated coefficients.
}
\description{
Extract Coefficients from an FQL Object
}
\examples{
data(quine, package = "MASS")
fit <- fql(Days ~ Eth + Sex + Age + Lrn, data = quine)
coef(fit)

}
