% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fql-methods.R
\name{confint.fql}
\alias{confint.fql}
\title{Confidence Intervals for FQL Coefficients}
\usage{
\method{confint}{fql}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{"fql"}.}

\item{parm}{A specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing,
all parameters are considered.}

\item{level}{The confidence level. Default is 0.95.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A matrix with columns for the lower and upper confidence limits.
}
\description{
Computes Wald-type confidence intervals for the regression coefficients.
}
\examples{
data(quine, package = "MASS")
fit <- fql(Days ~ Eth + Sex + Age + Lrn, data = quine)
confint(fit)
confint(fit, level = 0.99)

}
