% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ginv}
\alias{.ginv}
\title{Generalized Inverse of a Matrix}
\usage{
.ginv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{A numeric matrix.}

\item{tol}{Tolerance for singular value thresholding.}
}
\value{
The generalized inverse matrix.
}
\description{
Computes the Moore-Penrose generalized inverse using SVD.
}
\keyword{internal}
