% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.kernel_varest}
\alias{.kernel_varest}
\title{Kernel Variance Estimation}
\usage{
.kernel_varest(mi, mu, er, bw)
}
\arguments{
\item{mi}{The point at which to estimate variance.}

\item{mu}{Vector of all fitted means.}

\item{er}{Vector of squared residuals (y - mu)^2.}

\item{bw}{Bandwidth for the Gaussian kernel.}
}
\value{
Estimated variance at point mi.
}
\description{
Estimates the variance at a single point using a Gaussian kernel smoother
(local linear regression of squared residuals on fitted means).
}
\keyword{internal}
