% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.link_inv}
\alias{.link_inv}
\title{Apply Link Function (inverse)}
\usage{
.link_inv(eta)
}
\arguments{
\item{eta}{Linear predictor vector.}
}
\value{
Mean vector mu.
}
\description{
Applies the inverse link function to convert from linear predictor
to mean scale. Currently supports log link only.
}
\keyword{internal}
