% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fql-class.R
\name{.new_fql}
\alias{.new_fql}
\title{Construct an FQL Object}
\usage{
.new_fql(
  coefficients,
  se,
  p.value,
  vcov,
  fitted.values,
  residuals,
  variance,
  var.method.used,
  linear.predictors,
  converged,
  iterations,
  df.residual,
  n,
  formula,
  call,
  link,
  x,
  y,
  offset
)
}
\arguments{
\item{coefficients}{Named vector of regression coefficients.}

\item{se}{Named vector of standard errors.}

\item{p.value}{Named vector of p-values.}

\item{vcov}{Variance-covariance matrix.}

\item{fitted.values}{Fitted mean values.}

\item{residuals}{Response residuals.}

\item{variance}{Estimated variance function values.}

\item{var.method.used}{Character describing variance method used.}

\item{linear.predictors}{Linear predictor values.}

\item{converged}{Logical: did the algorithm converge?}

\item{iterations}{Number of outer iterations.}

\item{df.residual}{Residual degrees of freedom.}

\item{n}{Number of observations.}

\item{formula}{Model formula.}

\item{call}{Matched call.}

\item{link}{Link function name.}

\item{x}{Design matrix.}

\item{y}{Response vector.}

\item{offset}{Offset vector or NULL.}
}
\value{
An object of class \code{"fql"}.
}
\description{
Internal constructor for creating objects of class \code{"fql"}.
}
\keyword{internal}
