% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fql-methods.R
\name{fitted.fql}
\alias{fitted.fql}
\title{Extract Fitted Values from an FQL Object}
\usage{
\method{fitted}{fql}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fql"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A vector of fitted mean values.
}
\description{
Extract Fitted Values from an FQL Object
}
\examples{
data(quine, package = "MASS")
fit <- fql(Days ~ Eth + Sex + Age + Lrn, data = quine)
head(fitted(fit))

}
