% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fql-methods.R
\name{logLik.fql}
\alias{logLik.fql}
\title{Compute Quasi-Log-Likelihood for an FQL Object}
\usage{
\method{logLik}{fql}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fql"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An object of class \code{"logLik"} with the quasi-log-likelihood value.
}
\description{
Computes the quasi-log-likelihood value, defined as
\eqn{Q^*(\mu, y) = \sum_i \int_{y_i}^{\mu_i} (y_i - t) / V(t) dt}.
This is approximated numerically.
}
