% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fql-methods.R
\name{nobs.fql}
\alias{nobs.fql}
\title{Extract the Number of Observations from an FQL Object}
\usage{
\method{nobs}{fql}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fql"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Integer: the number of observations.
}
\description{
Extract the Number of Observations from an FQL Object
}
