% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fql-methods.R
\name{print.summary.fql}
\alias{print.summary.fql}
\title{Print Summary of an FQL Object}
\usage{
\method{print}{summary.fql}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars", TRUE),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"summary.fql"}.}

\item{digits}{Number of significant digits.}

\item{signif.stars}{Logical: print significance stars?}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Summary of an FQL Object
}
