% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fql-methods.R
\name{residuals.fql}
\alias{residuals.fql}
\title{Extract Residuals from an FQL Object}
\usage{
\method{residuals}{fql}(object, type = c("response", "pearson", "working"), ...)
}
\arguments{
\item{object}{An object of class \code{"fql"}.}

\item{type}{The type of residuals: \code{"response"} (default, y - mu),
\code{"pearson"} ((y - mu) / sqrt(V(mu))), or \code{"working"}
((y - mu) / mu for log link).}

\item{...}{Additional arguments (currently unused).}
}
\value{
A vector of residuals.
}
\description{
Extract Residuals from an FQL Object
}
\examples{
data(quine, package = "MASS")
fit <- fql(Days ~ Eth + Sex + Age + Lrn, data = quine)
head(residuals(fit, type = "pearson"))

}
