% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fql-methods.R
\name{vcov.fql}
\alias{vcov.fql}
\title{Extract Variance-Covariance Matrix from an FQL Object}
\usage{
\method{vcov}{fql}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fql"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The variance-covariance matrix.
}
\description{
Returns the sandwich variance-covariance matrix of the estimated
regression coefficients.
}
\examples{
data(quine, package = "MASS")
fit <- fql(Days ~ Eth + Sex + Age + Lrn, data = quine)
vcov(fit)

}
